pyproject.toml
[tool.poetry]
name = "eva"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.7"
Django = "^3.2.9"
Pillow = "^9.0.1"
gunicorn = "^20.1.0"
whitenoise = "^6.0.0"
psycopg2-binary = "^2.9.3"
psycopg2 = "2.7.5"
python-dotenv = "^0.20.0"
dj-database-url = "^0.5.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
